openlearn_django/requirement.txt
asgiref==3.8.1
asttokens==3.0.0
certifi==2024.2.2
charset-normalizer==3.3.2
colorama==0.4.6
comm==0.2.2
contourpy==1.3.1
cycler==0.12.1
debugpy==1.8.13
decorator==5.2.1
Django==5.0.4
django-cors-headers==4.3.1
django-extensions==4.1
django-filter==24.2
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
executing==2.2.0
fonttools==4.56.0
future==1.0.0
idna==3.7
ipykernel==6.29.5
ipython==9.0.2
ipython_pygments_lexers==1.1.1
jedi==0.19.2
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.8
Markdown==3.6
matplotlib==3.10.1
matplotlib-inline==0.1.7
mizani==0.13.1
nest-asyncio==1.6.0
numpy==2.2.3
opencv-python==4.11.0.86
packaging==24.2
pandas==2.2.3
parso==0.8.4
patsy==1.0.1
pillow==11.1.0
platformdirs==4.3.6
plotnine==0.14.5
prompt_toolkit==3.0.50
psutil==7.0.0
pure_eval==0.2.3
pygame==2.6.1
pygame-widgets==1.1.7
Pygments==2.19.1
PyJWT==2.8.0
pyparsing==3.2.1
pytesseract==0.3.13
python-dateutil==2.9.0.post0
pytz==2025.1
pywin32==309
pyzmq==26.3.0
requests==2.31.0
scipy==1.15.2
screeninfo==0.8.1
serializers==0.2.4
six==1.17.0
sqlparse==0.5.0
stack-data==0.6.3
statsmodels==0.14.4
tesseract==0.1.3
tornado==6.4.2
traitlets==5.14.3
tzdata==2025.1
urllib3==2.2.1
wcwidth==0.2.13
wget==3.2
whitenoise==6.9.0
